requirements.txt
anthropic>=0.39
PyMuPDF>=1.24
Pillow>=10.0
pytesseract>=0.3.10
python-docx>=1.1
